+++ NEWS.md
@@ -1,4 +1,5 @@
 ## v3.0.0 (IN PROGRESS)
+- Base64Url hash method introduced [MODLD-1087](https://folio-org.atlassian.net/browse/MODLD-1087)
 
 ## v2.0.0 (17-04-2026)
 - Add LIGHT_RESOURCE type [MODLD-980](https://folio-org.atlassian.net/browse/MODLD-980)